Fixed utility dimension loading issues
This commit should fix a longstanding issue with the utility dimension
not loading on the first load of the world.

This has involved modifying the `/loadchunk` command in the Plagued
Skies mod to allow for chunkloading after a delay, allowing for the
dimension to fully spawn before trying to chunkload it.

This also involved updating the defaultworldgenerator config files to
use the new `/loadchunk` command, upading the mod list (to reflect the
updated Plagued Skies mod), and bumping the version number.

This branch will be the beginning of the next Creeping Dark update,
which will involve updating all the involved mods, as well as tweaks
made in the 1.6.1.1 branch.